// src/models/styles.ts
export interface ContainerProps {
  width?: string;
  height?: string;
  flex?: boolean;
  column?: string;
  justify?: string;
  alignItems?: string;
  gap?: string;
  padding?: string;
  margin?: string;
}

export interface ButtonProps {
  padding?: string;
  width?: string;
}

export interface SpanProps {
  completed: boolean;
}